requirements.txt
torch
opencv-python
imageio
Pillow
numpy
ultralytics
